Revert "Contributing changes back to source project"
