+++ Enums.cs
@@ -1,5 +1,4 @@
-﻿using iTextSharp.text;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -34,12 +33,7 @@ namespace SharpPDFLabel
             UNDERLINE = 4,
         }
 
-        public enum Alignment
-        {
-            LEFT = Element.ALIGN_LEFT,
-            CENTER = Element.ALIGN_CENTER,
-            RIGHT = Element.ALIGN_RIGHT
-        }
+        
 
     }
 }
