+++ Label.cs
@@ -1,96 +1,116 @@
-﻿using iTextSharp.text;
-using iTextSharp.text.pdf;
-using System;
+﻿using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using System.Text;
 
 namespace SharpPDFLabel
 {
-    public class Label
+    /// <summary>
+    /// The base class for a Label, all the actual label definition classes derive from this
+    /// Dimensions and margins are defined in mm but converted to Points when read back (for use in iTextSharp)
+    /// </summary>
+    public abstract class Label
     {
-        private List<byte[]> _images;
-        private List<TextChunk> _textChunks;
-        private Enums.Alignment _hAlign;
-
-        /// sets align to CENTER
-        public Label()
-            : this(Enums.Alignment.CENTER)
+        /// <summary>
+        /// The width of 1 label
+        /// </summary>
+        protected double _Width;
+        public float Width
         {
-        }
-
-        /// <param name="hAlign">horizontal alignment: LEFT, CENTER, RIGHT</param>
-        public Label(Enums.Alignment hAlign)
-        {
-            _images = new List<byte[]>();
-            _textChunks = new List<TextChunk>();
-            _hAlign = hAlign;
-        }
-
-
-        public PdfPCell GetLabelCell()
-        {
-            // Create a new Phrase and add the image to it
-            var cellContent = new Phrase();
-
-            foreach (var img in _images)
-            {
-                var pdfImg = iTextSharp.text.Image.GetInstance(img);
-                cellContent.Add(new Chunk(pdfImg, 0, 0));
-            }
-
-            foreach (var txt in _textChunks)
-            {
-                var font = FontFactory.GetFont(txt.FontName, BaseFont.CP1250, txt.EmbedFont, txt.FontSize, txt.FontStyle);
-                cellContent.Add(new Chunk("\n" + txt.Text, font));
-            }
-
-            //Create a new cell specifying the content
-            var cell = new PdfPCell(cellContent);
-            cell.HorizontalAlignment = (int)_hAlign;
-            cell.VerticalAlignment = Element.ALIGN_TOP;
-
-            return cell;
-        }
-
-
-        private void CopyStream(Stream input, Stream output)
-        {
-            byte[] b = new byte[32768];
-            int r;
-            while ((r = input.Read(b, 0, b.Length)) > 0)
-                output.Write(b, 0, r);
+            get { return mmToPoint(_Width); }
         }
 
         /// <summary>
-        /// Add an image to the labels
-        /// Currently adds images and then text in that specific order
+        /// The height of 1 label
         /// </summary>
-        /// <param name="img"></param>
-        public void AddImage(Stream img)
+        protected double _Height;
+        public float Height
         {
+            get { return mmToPoint(_Height); }
         }
-        /// <summary>
-        /// Add a chunk of text to the labels
-        /// </summary>
-        /// <param name="text">The text to add e.g "I am on a label"</param>
-        /// <param name="fontName">The name of the font e.g. "Verdana"</param>
-        /// <param name="fontSize">The font size in points e.g. 12</param>
-        /// <param name="embedFont">If the font you are using may not be on the target machine, set this to true</param>
-        /// <param name="fontStyles">An array of required font styles</param>
-        public void AddText(string text, string fontName, int fontSize, bool embedFont = false, params Enums.FontStyle[] fontStyles)
-        {
-            int fontStyle = 0;
-            if (fontStyles != null)
-            {
-                foreach (var item in fontStyles)
-                {
-                    fontStyle += (int)item;
-                }
-            }
 
-            _textChunks.Add(new TextChunk() { Text = text, FontName = fontName, FontSize = fontSize, FontStyle = fontStyle, EmbedFont = embedFont });
+        /// <summary>
+        /// The width of the horizontal gap between labels
+        /// </summary>
+        protected double _HorizontalGapWidth;
+        public float HorizontalGapWidth
+        {
+            get { return mmToPoint(_HorizontalGapWidth); }
+        }
+
+        /// <summary>
+        /// The height of the vertical gap between labels
+        /// </summary>
+        protected double _VerticalGapHeight;
+        public float VerticalGapHeight
+        {
+            get { return mmToPoint(_VerticalGapHeight); }
+        }
+
+        /// <summary>
+        /// The left page margin
+        /// </summary>
+        protected double _PageMarginLeft;
+        public float PageMarginLeft
+        {
+            get { return mmToPoint(_PageMarginLeft); }
+        }
+
+        /// <summary>
+        /// The right page margin
+        /// </summary>
+        protected double _PageMarginRight;
+        public float PageMarginRight
+        {
+            get { return mmToPoint(_PageMarginRight); }
+        }
+
+        /// <summary>
+        /// The top page margin
+        /// </summary>
+        protected double _PageMarginTop;
+        public float PageMarginTop
+        {
+            get { return mmToPoint(_PageMarginTop); }
+        }
+
+        /// <summary>
+        /// The bottom page margin
+        /// </summary>
+        protected double _PageMarginBottom;
+        public float PageMarginBottom
+        {
+            get { return mmToPoint(_PageMarginBottom); }
+        }
+
+        /* page definitions */
+        
+        /// <summary>
+        /// The paper size
+        /// </summary>
+        public Enums.PageSize PageSize { get; set; }
+        
+        
+        /// <summary>
+        /// The number of labels running across the page
+        /// </summary>
+        public int LabelsPerRow { get; set; }
+        
+        
+        /// <summary>
+        /// The number of labels running down the page
+        /// </summary>
+        public int LabelRowsPerPage { get; set; }
+
+
+        /// <summary>
+        /// iTextSharp uses points as its units
+        /// </summary>
+        /// <param name="mm">Millimetres to convert</param>
+        /// <returns>millimetres converted to points represented by a float</returns>
+        private float mmToPoint(double mm)
+        {
+            return (float)((mm / 25.4) * 72);
         }
 
     }
