+++ packages.config
@@ -1,4 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="iTextSharp" version="5.5.7" targetFramework="net40" />
+  <package id="iTextSharp" version="5.1.1" />
 </packages>
